+++ domutils/legs/col_utils/tests/test_col_utils.py
@@ -1,5 +1,5 @@
 import unittest
-import col_utils
+import domutils.legs.col_utils as col_utils
 import numpy as np
 
 
@@ -22,6 +22,24 @@ class TestStringMethods(unittest.TestCase):
         self.assertEqual(col_utils.col_pair(col_txt).tolist(), [[[169,222,255],[0,   81,237]],
                                                                 [[255,190,187],[158,  0, 13]],
                                                                 [[255,255,255],[  0,  0,  0]]])
+
+    def test_col_rgb_wrong_col(self):
+        with self.assertRaises(ValueError):
+            col_utils.col_rgb('non_existing_col')
+
+    def test_col_rgb_partial_match(self):
+        with self.assertRaises(ValueError):
+            col_utils.col_rgb('dark_')
+
+    def test_col_rgb_good_col(self):
+        self.assertEqual(col_utils.col_rgb('white').tolist(), [255,255,255])
+
+    def test_col_rgb_good_col_multi(self):
+        self.assertEqual(col_utils.col_rgb(['white', 'black']).tolist(), [[255,255,255],[0,0,0]])
+
+    def test_col_rgb_bad_col_multi(self):
+        with self.assertRaises(ValueError):
+            map0 = col_utils.col_rgb(['white', 'bad'])
 
 if __name__ == '__main__':
     unittest.main()