+++ domutils/legs/validation_tools/rgb.py
@@ -4,18 +4,18 @@ def rgb(col_rgb, var_name=None, instructions=None):
     import numpy as np
     from os import linesep as newline
     try: 
-        col_rgb_flt = np.asarray(col_rgb)
+        col_rgb_int16 = np.asarray(col_rgb, dtype=np.int16)
     except:
         err_mess      =  newline+' '                               + newline
         if var_name is not None:
             err_mess +=('Problem with the keyword "'+var_name+'"'   + newline)
-        err_mess     +=('Provided values must be convertible to a numpy array.')
+        err_mess     +=('Provided values must be convertible to a numpy integer array.')
         if instructions is not None:
             err_mess += instructions
         raise TypeError(err_mess)
 
     #check range of values
-    if (col_rgb_flt < 0).any() or (col_rgb_flt > 255).any():
+    if (col_rgb_int16 < 0).any() or (col_rgb_int16 > 255).any():
         err_mess      =  newline+' '                               + newline
         if var_name is not None:
             err_mess +=('Problem with the keyword "'+var_name+'"'   + newline)
@@ -25,7 +25,7 @@ def rgb(col_rgb, var_name=None, instructions=None):
         raise ValueError(err_mess)
 
     #RGB values should come in three
-    if np.remainder(col_rgb_flt.size, 3) != 0:
+    if np.remainder(col_rgb_int16.size, 3) != 0:
         err_mess      =  newline+' '                               + newline
         if var_name is not None:
             err_mess +=('Problem with the keyword "'+var_name+'"'   + newline)
@@ -34,4 +34,4 @@ def rgb(col_rgb, var_name=None, instructions=None):
             err_mess += instructions
         raise ValueError(err_mess)
 
-    return col_rgb_flt.astype('uint8')
+    return col_rgb_int16.astype('uint8')