+++ domutils/legs/validation_tools/tests/test_n_col.py
@@ -1,6 +1,5 @@
 import unittest
-import col_utils
-import validation_tools as validate
+import domutils.legs.validation_tools as validate
 
 
 class TestStringMethods(unittest.TestCase):