+++ domutils/legs/validation_tools/tests/test_not_mapped_twice.py
@@ -1,6 +1,6 @@
 import unittest
 import numpy as np
-import validation_tools as validate
+import domutils.legs.validation_tools as validate
 
 
 class TestStringMethods(unittest.TestCase):